refactor(kube): move the whole kube.service impl set to .internal (#546)
Completes the jhelm-kube half of the internal-package migration. Every
implementation type in org.alexmond.jhelm.kube.service is a detail wired by
the module auto-config or reached through the core KubeService interface —
none is referenced outside jhelm-kube, and KubeClient's key accessor
(apiClient()) is package-private, so the whole set is internal by design.

Moved into ...kube.service.internal (joining the decorators from #601):
KubeClient, KubernetesClientProvider, KubernetesHealthIndicator,
HelmKubeService, AsyncHelmKubeService, ResourcePluralizer. Keeping them in
one package preserves KubeClient's package-private apiClient() access. The
auto-config + its test import from the new package; the higher-level flow/
lifecycle integration tests stay at the service level and import
HelmKubeService from internal.

jhelm-kube's public surface is now just the auto-config, properties, and the
core KubeService interface. jhelm-kube + jhelm-cli build green (verified
against cli, which lacks actuator). Next: core's RepoHttpClientFactory /
static factories, then the per-module "what's stable" statement.
