+++ jhelm-kube/src/main/java/org/alexmond/jhelm/kube/JhelmKubeAutoConfiguration.java
@@ -11,9 +11,9 @@ import org.alexmond.jhelm.core.JhelmMetricsAutoConfiguration;
 import org.alexmond.jhelm.core.metrics.JhelmMetrics;
 import org.alexmond.jhelm.core.service.KubeService;
 import org.alexmond.jhelm.kube.config.JhelmKubernetesProperties;
-import org.alexmond.jhelm.kube.service.AsyncHelmKubeService;
-import org.alexmond.jhelm.kube.service.KubeClient;
-import org.alexmond.jhelm.kube.service.KubernetesHealthIndicator;
+import org.alexmond.jhelm.kube.service.internal.AsyncHelmKubeService;
+import org.alexmond.jhelm.kube.service.internal.KubeClient;
+import org.alexmond.jhelm.kube.service.internal.KubernetesHealthIndicator;
 import org.alexmond.jhelm.kube.service.internal.ObservableKubeService;
 import org.alexmond.jhelm.kube.service.internal.RetryableKubeService;
 import org.springframework.beans.factory.ObjectProvider;
