+++ jhelm-kube/src/test/java/org/alexmond/jhelm/kube/JhelmKubeAutoConfigurationTest.java
@@ -5,8 +5,8 @@ import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import org.alexmond.jhelm.core.JhelmMetricsAutoConfiguration;
 import org.alexmond.jhelm.core.metrics.JhelmMetrics;
 import org.alexmond.jhelm.core.service.KubeService;
-import org.alexmond.jhelm.kube.service.HelmKubeService;
-import org.alexmond.jhelm.kube.service.KubeClient;
+import org.alexmond.jhelm.kube.service.internal.HelmKubeService;
+import org.alexmond.jhelm.kube.service.internal.KubeClient;
 import org.alexmond.jhelm.kube.service.internal.ObservableKubeService;
 import org.alexmond.jhelm.kube.service.internal.RetryableKubeService;
 import org.junit.jupiter.api.Test;
