+++ jhelm-kube/src/test/java/org/alexmond/jhelm/kube/service/HelmFullFlowIntegrationTest.java
@@ -1,5 +1,7 @@
 package org.alexmond.jhelm.kube.service;
 
+import org.alexmond.jhelm.kube.service.internal.HelmKubeService;
+
 import lombok.extern.slf4j.Slf4j;
 import org.alexmond.jhelm.core.model.Chart;
 import org.alexmond.jhelm.core.model.ChartMetadata;
