+++ jhelm-kube/src/test/java/org/alexmond/jhelm/kube/service/internal/KubernetesHealthIndicatorTest.java
@@ -1,4 +1,4 @@
-package org.alexmond.jhelm.kube.service;
+package org.alexmond.jhelm.kube.service.internal;
 
 import io.kubernetes.client.openapi.ApiClient;
 import org.junit.jupiter.api.Test;
